Couple of tweaks
 * Use set -e in build.sh so the script terminate on any error
 * Fix integer width in owned_packet.rs
 * Remove auto-formatting settings from .vscode. Let users set them as
   they prefer.

+++ libconntrack/src/owned_packet.rs
@@ -558,7 +558,7 @@ impl<'de> Deserialize<'de> for OwnedParsedPacket {
                                 map.next_value::<(i64, i64, u32, u32)>()?;
                             pkt.pcap_header = PacketHeader {
                                 ts: libc::timeval {
-                                    tv_sec: sec as i32,
+                                    tv_sec: sec,
                                     tv_usec: usec as i32,
                                 },
                                 caplen,
